Shop Updates
You can buy items and the buy button will update visually and the database will store it. On the page being refreshed, visually it is set after a buffer moment to being bought. Currently it is set to being the id value of the item but that can be changed.