added savings option
just bare minimum concept, in future i want this to be an output and have a simple input that identifies amount spent and what category it fits into. that allows the savings to be an output of their spent allowance. an edit button could be implemented to make changes to how much is budgeted