swapped url to cookies
user gets wiped from window when swapping between signup and the main page, to retain the small user info i was using a url which is insecure and didnt work for actual sites. Using cookies doesnt have the information easily obtainable in the url (and it actually works now). i think